Separate materialising an upstream PR from grading one
Turning a candidate row into a diff on disk is the only part of the
upstream-PR lane that is genuinely its own. Everything after it -- apply,
structural vetting, bench, KEEP/REVERT, revert, KB record -- is what every
other patch source already does, and ``framework_agent`` carries a second
copy of it that has already drifted from ``integrate_patch``'s.

Move the materialisation half to ``_patch_source_pr``: the three sources in
priority order (explicit paths, a net diff from a worktree at the PR head,
the raw diff_url) plus the candidate/repo/HEAD helpers they need. Pure
move, no behaviour change; ``framework_agent`` imports what it used to
define and drops from 1269 to 1008 lines.

This is the seam the patch-source strategy lands on next.
